feat(test): Consolidate dialog tests into comprehensive test suite
• **Merged Testing Strategy**: Combined Level 2 UAT and Level 3 destructive testing into single comprehensive file
• **Meaningful Organization**: Restructured tests by dialog type rather than testing level for better maintainability
• **Enhanced Security Coverage**: Comprehensive vulnerability testing with proper XFAIL documentation for 2 security issues
• **Quality Compliance**: Full adherence to ruff, pylint standards with acceptable mypy patterns consistent with codebase

**Test Organization by Dialog Type:**
- TestBaseDialogFunctionality: ArtisanDialog core features + destructive testing (6 tests)
- TestMessageBoxFunctionality: ArtisanMessageBox timeout and feedback (2 tests)
- TestHelpDialogFunctionality: HelpDlg search + security testing (9 tests)
- TestInputDialogFunctionality: ArtisanInputDialog + injection testing (4 tests)
- TestSelectionDialogFunctionality: ComboBox selection workflows (2 tests)

**Key Achievements:**
- 21 total tests: 19 passed, 2 xfailed (documented vulnerabilities)
- Hypothesis property-based testing with health check management
- Comprehensive security testing covering Levels 1-4 protection
- Single source of truth for all dialog testing with shared fixtures
- Vulnerability documentation with remediation guidance

**Security Findings:**
- XFAIL: Memory exhaustion in search functionality (needs match limiting)
- XFAIL: No URL validation in drag-and-drop (needs input sanitization)
- VALIDATED: Regex injection protection working correctly

Files: test_dialogs.py (consolidated), removed test_dialogs_destructive.py
