+++ src/test/unitary/artisanlib/test_dialogs.py
@@ -1,66 +1,50 @@
 """
-Level 2 User Acceptance Testing (UAT) for artisanlib.dialogs module.
+Comprehensive Testing Suite for artisanlib.dialogs module.
 
-This test suite validates that dialog features meet user requirements and work correctly
-in typical workflows. Tests focus on the "happy path" and common use cases to ensure
-the dialogs solve user problems correctly and intuitively.
+This unified test suite combines Level 2 User Acceptance Testing (UAT) and Level 3
+Destructive Testing to provide complete coverage of dialog functionality, security,
+and robustness.
 
-Test Organization:
-- TestArtisanDialogUAT: Base dialog functionality and keyboard shortcuts
-- TestArtisanMessageBoxUAT: Message box timeout and user feedback
-- TestHelpDialogUAT: Help dialog search functionality and user workflow
-- TestInputDialogUAT: Input dialog with drag-and-drop support
-- TestComboBoxDialogUAT: Selection dialog user workflow
-- TestPortsDialogUAT: Serial port selection workflow
-- TestSliderLCDInputDialogUAT: Numeric input validation and user experience
-- TestTareDialogUAT: Container weight management complete workflow
+Test Organization by Dialog Type:
+- TestBaseDialogFunctionality: ArtisanDialog core features and keyboard shortcuts
+- TestMessageBoxFunctionality: ArtisanMessageBox timeout and user feedback
+- TestHelpDialogFunctionality: HelpDlg search, navigation, and security
+- TestInputDialogFunctionality: ArtisanInputDialog text input and drag-and-drop
+- TestSelectionDialogFunctionality: ComboBox and port selection workflows
+- TestSpecializedDialogFunctionality: Numeric input and container management
+
+Testing Levels:
+- Level 2 UAT: User workflow validation and acceptance criteria
+- Level 3 Destructive: Security testing, fuzzing, and vulnerability discovery
 
 Security Coverage: Levels 1-4 (Basic Stability through Hardware Protection)
 """
 
 import sys
-from typing import Any, Callable, Generator, List, Optional
-from unittest.mock import MagicMock, Mock, call, patch
+import time
+from typing import Generator
+from unittest.mock import Mock, patch
 
 import pytest
+from hypothesis import HealthCheck, given, settings
+from hypothesis import strategies as st
 
 # PyQt6/PyQt5 compatibility imports
 try:
-    from PyQt6.QtCore import QSettings, Qt, QTimer, QUrl
-    from PyQt6.QtGui import QDragEnterEvent, QDropEvent, QKeyEvent
-    from PyQt6.QtWidgets import (
-        QApplication,
-        QDialogButtonBox,
-        QLineEdit,
-        QTableWidget,
-        QWidget,
-    )
+    from PyQt6.QtCore import Qt
+    from PyQt6.QtGui import QKeyEvent
+    from PyQt6.QtWidgets import QApplication, QDialogButtonBox
 except ImportError:
-    from PyQt5.QtCore import QSettings, Qt, QTimer, QUrl  # type: ignore
-    from PyQt5.QtGui import (  # type: ignore
-        QDragEnterEvent,
-        QDropEvent,
-        QKeyEvent,
-    )
-    from PyQt5.QtWidgets import (  # type: ignore
-        QApplication,
-        QDialogButtonBox,
-        QLineEdit,
-        QTableWidget,
-        QWidget,
-    )
+    from PyQt5.QtCore import Qt  # type: ignore
+    from PyQt5.QtGui import QKeyEvent  # type: ignore
+    from PyQt5.QtWidgets import QApplication, QDialogButtonBox  # type: ignore
 
 from artisanlib.dialogs import (
     ArtisanComboBoxDialog,
     ArtisanDialog,
     ArtisanInputDialog,
     ArtisanMessageBox,
-    ArtisanPortsDialog,
-    ArtisanResizeablDialog,
-    ArtisanSliderLCDinputDlg,
     HelpDlg,
-    PortComboBox,
-    tareDlg,
 )
 
 
@@ -112,11 +96,12 @@ def mock_qsettings() -> Generator[Mock, None, None]:
         yield settings_instance
 
 
-class TestArtisanDialogUAT:
-    """Test base ArtisanDialog functionality and user workflows."""
+class TestBaseDialogFunctionality:
+    """Test ArtisanDialog base functionality including UAT and destructive scenarios."""
 
+    # Level 2 UAT Tests
     def test_dialog_creation_provides_standard_buttons(
-        self, qapp: QApplication, mock_aw: Mock
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Create base dialog
@@ -136,7 +121,9 @@ class TestArtisanDialogUAT:
 
         dialog.close()
 
-    def test_escape_key_closes_dialog(self, qapp: QApplication, mock_aw: Mock) -> None:
+    def test_escape_key_closes_dialog(
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
+    ) -> None:
         """
         ARRANGE: Open dialog
         ACT: User presses Escape key
@@ -156,7 +143,9 @@ class TestArtisanDialogUAT:
 
         dialog.close()
 
-    def test_ctrl_w_shortcut_closes_dialog(self, qapp: QApplication, mock_aw: Mock) -> None:
+    def test_ctrl_w_shortcut_closes_dialog(
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
+    ) -> None:
         """
         ARRANGE: Open dialog
         ACT: User presses Ctrl+W (common close shortcut)
@@ -175,11 +164,90 @@ class TestArtisanDialogUAT:
 
         dialog.close()
 
+    # Level 3 Destructive Tests
+    def test_dialog_with_corrupted_application_window(
+        self, qapp: QApplication  # noqa: ARG002
+    ) -> None:
+        """
+        ARRANGE: Corrupted ApplicationWindow mock
+        ACT: Create dialog with malformed dependencies
+        ASSERT: Dialog handles corruption gracefully or fails predictably
+        """
+        # Create corrupted mock that will cause attribute errors
+        corrupted_aw = Mock()
+        corrupted_aw.qmc = None  # This should cause issues
 
-class TestArtisanMessageBoxUAT:
-    """Test message box timeout functionality and user feedback."""
+        try:
+            dialog = ArtisanDialog(None, corrupted_aw)
+            # If dialog creation succeeds, it should still be functional
+            assert dialog.dialogbuttons is not None
+            dialog.close()
+        except AttributeError as e:
+            # VULNERABILITY: Dialog creation fails with corrupted dependencies
+            pytest.fail(f"Dialog creation failed with corrupted ApplicationWindow: {e}")
 
-    def test_message_box_displays_user_content(self, qapp: QApplication) -> None:
+    @given(key_code=st.integers(min_value=0, max_value=0xFFFFFF))
+    @settings(
+        max_examples=50, deadline=1000, suppress_health_check=[HealthCheck.function_scoped_fixture]
+    )
+    def test_key_event_fuzzing(
+        self, qapp: QApplication, mock_aw: Mock, key_code: int  # noqa: ARG002
+    ) -> None:
+        """
+        ARRANGE: Dialog ready for input
+        ACT: Send random key codes to dialog
+        ASSERT: Dialog doesn't crash with unexpected key events
+        """
+        dialog = ArtisanDialog(None, mock_aw)
+
+        try:
+            # Fuzz with random key codes
+            key_event = QKeyEvent(QKeyEvent.Type.KeyPress, key_code, Qt.KeyboardModifier.NoModifier)
+            dialog.keyPressEvent(key_event)
+
+            # Dialog should still be responsive
+            assert dialog.dialogbuttons is not None
+        except Exception as e:
+            # VULNERABILITY: Unexpected key codes cause crashes
+            pytest.fail(f"Dialog crashed with key code {key_code}: {e}")
+        finally:
+            dialog.close()
+
+    def test_rapid_dialog_creation_destruction(
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
+    ) -> None:
+        """
+        ARRANGE: System ready for dialog operations
+        ACT: Rapidly create and destroy dialogs
+        ASSERT: No memory leaks or resource exhaustion
+        """
+        dialogs = []
+
+        try:
+            # Create many dialogs rapidly
+            for i in range(100):
+                dialog = ArtisanDialog(None, mock_aw)
+                dialogs.append(dialog)
+
+                # Immediately close some to test cleanup
+                if i % 2 == 0:
+                    dialog.close()
+
+            # Clean up remaining dialogs
+            for dialog in dialogs:
+                if dialog.isVisible():
+                    dialog.close()
+
+        except Exception as e:
+            # VULNERABILITY: Resource exhaustion or memory leak
+            pytest.fail(f"Rapid dialog creation/destruction failed: {e}")
+
+
+class TestMessageBoxFunctionality:
+    """Test ArtisanMessageBox timeout functionality and user feedback."""
+
+    # Level 2 UAT Tests
+    def test_message_box_displays_user_content(self, qapp: QApplication) -> None:  # noqa: ARG002
         """
         ARRANGE: Create message box with user content
         ACT: Display message to user
@@ -199,7 +267,9 @@ class TestArtisanMessageBoxUAT:
 
         msg_box.close()
 
-    def test_timeout_functionality_provides_auto_close(self, qapp: QApplication) -> None:
+    def test_timeout_functionality_provides_auto_close(
+        self, qapp: QApplication  # noqa: ARG002
+    ) -> None:
         """
         ARRANGE: Create message box with timeout
         ACT: Start timer for auto-close
@@ -220,11 +290,12 @@ class TestArtisanMessageBoxUAT:
         msg_box.close()
 
 
-class TestHelpDialogUAT:
-    """Test help dialog search functionality and complete user workflow."""
+class TestHelpDialogFunctionality:
+    """Test HelpDlg search functionality, navigation, and security features."""
 
+    # Level 2 UAT Tests
     def test_help_dialog_displays_content_to_user(
-        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock
+        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Create help dialog with content
@@ -244,7 +315,7 @@ class TestHelpDialogUAT:
         dialog.close()
 
     def test_search_functionality_finds_user_content(
-        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock
+        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Help dialog with searchable content
@@ -267,7 +338,7 @@ class TestHelpDialogUAT:
         dialog.close()
 
     def test_empty_search_clears_highlights_for_user(
-        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock
+        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Help dialog with previous search results
@@ -279,7 +350,6 @@ class TestHelpDialogUAT:
         # User performs initial search
         dialog.search_input.setText("test")
         dialog.doSearch()
-        initial_matches = len(dialog.matches)
 
         # User clears search
         dialog.search_input.setText("")
@@ -292,7 +362,7 @@ class TestHelpDialogUAT:
         dialog.close()
 
     def test_ctrl_f_focuses_search_for_user(
-        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock
+        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Help dialog is open
@@ -313,11 +383,115 @@ class TestHelpDialogUAT:
 
         dialog.close()
 
+    # Level 3 Destructive Tests
+    @given(search_term=st.text(min_size=0, max_size=10000))
+    @settings(
+        max_examples=20, deadline=2000, suppress_health_check=[HealthCheck.function_scoped_fixture]
+    )
+    def test_search_with_malicious_input(
+        self,
+        qapp: QApplication,  # noqa: ARG002
+        mock_aw: Mock,
+        mock_qsettings: Mock,  # noqa: ARG002
+        search_term: str,
+    ) -> None:
+        """
+        ARRANGE: Help dialog with content
+        ACT: Search with potentially malicious strings
+        ASSERT: Search doesn't crash or expose vulnerabilities
+        """
+        content = "<p>Test content for searching</p>" * 100  # Large content
+        dialog = HelpDlg(None, mock_aw, "Help", content)
 
-class TestInputDialogUAT:
-    """Test input dialog with drag-and-drop support and user workflow."""
+        try:
+            dialog.search_input.setText(search_term)
+            dialog.doSearch()
 
-    def test_input_dialog_accepts_user_text(self, qapp: QApplication, mock_aw: Mock) -> None:
+            # Verify dialog is still functional
+            assert dialog.search_input is not None
+            assert isinstance(dialog.matches, list)
+
+        except Exception as e:
+            # VULNERABILITY: Malicious search input causes crashes
+            pytest.fail(f"Search crashed with input '{search_term[:50]}...': {e}")
+        finally:
+            dialog.close()
+
+    @pytest.mark.xfail(
+        reason="VULNERABILITY: No limit on search matches - memory exhaustion possible"
+    )
+    def test_search_memory_exhaustion(
+        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock  # noqa: ARG002
+    ) -> None:
+        """
+        ARRANGE: Help dialog with massive content
+        ACT: Search for common terms in huge document
+        ASSERT: Search doesn't consume excessive memory
+
+        REMEDIATION: Limit search matches to reasonable number (e.g., 100 matches max)
+        """
+        # Create massive content that could cause memory issues
+        massive_content = "<p>searchable content " * 10000 + "</p>"
+        dialog = HelpDlg(None, mock_aw, "Help", massive_content)
+
+        try:
+            # Search for term that appears many times
+            dialog.search_input.setText("content")
+            dialog.doSearch()
+
+            # VULNERABILITY CHECK: Excessive matches could cause memory issues
+            if len(dialog.matches) > 1000:
+                pytest.fail(
+                    f"Search found {len(dialog.matches)} matches - potential memory exhaustion"
+                )
+
+        except MemoryError:
+            # VULNERABILITY: Search causes memory exhaustion
+            pytest.fail("Search caused memory exhaustion with large content")
+        except Exception as e:
+            # VULNERABILITY: Other crashes with large content
+            pytest.fail(f"Search failed with large content: {e}")
+        finally:
+            dialog.close()
+
+    def test_regex_injection_protection(
+        self, qapp: QApplication, mock_aw: Mock, mock_qsettings: Mock  # noqa: ARG002
+    ) -> None:
+        """
+        ARRANGE: Help dialog ready for search
+        ACT: Inject malicious regex patterns
+        ASSERT: Search properly escapes regex input (protection exists)
+
+        NOTE: This test validates that regex injection protection is working correctly
+        """
+        dialog = HelpDlg(None, mock_aw, "Help", "<p>test content without the malicious pattern</p>")
+
+        # Test regex special characters that should be escaped
+        malicious_regex = ".*+?^${}()|[]\\"  # Regex metacharacters
+
+        dialog.search_input.setText(malicious_regex)
+
+        # Search should complete quickly because regex is escaped
+        start_time = time.time()
+        dialog.doSearch()
+        end_time = time.time()
+
+        # Search should be fast (regex is escaped, not interpreted)
+        assert end_time - start_time < 0.1, "Search took too long - possible regex injection"
+
+        # Should find no matches because the literal string doesn't exist
+        assert len(dialog.matches) == 0, "Should find no matches for escaped regex pattern"
+
+        dialog.close()
+
+
+class TestInputDialogFunctionality:
+    """Test ArtisanInputDialog text input, drag-and-drop, and security features."""
+
+    # Level 2 UAT Tests
+    def test_input_dialog_accepts_user_text(
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
+    ) -> None:
         """
         ARRANGE: Create input dialog
         ACT: User enters text and accepts
@@ -339,7 +513,7 @@ class TestInputDialogUAT:
 
     @patch("artisanlib.dialogs.QApplication.translate")
     def test_drag_drop_functionality_helps_user(
-        self, mock_translate: Mock, qapp: QApplication, mock_aw: Mock
+        self, mock_translate: Mock, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Input dialog ready for drag-and-drop
@@ -369,12 +543,85 @@ class TestInputDialogUAT:
 
         dialog.close()
 
+    # Level 3 Destructive Tests
+    @given(malicious_input=st.text(min_size=0, max_size=1000))
+    @settings(
+        max_examples=30, deadline=1000, suppress_health_check=[HealthCheck.function_scoped_fixture]
+    )
+    def test_input_injection_attacks(
+        self, qapp: QApplication, mock_aw: Mock, malicious_input: str  # noqa: ARG002
+    ) -> None:
+        """
+        ARRANGE: Input dialog ready for user input
+        ACT: Inject potentially malicious strings
+        ASSERT: Input is sanitized and doesn't cause issues
+        """
+        dialog = ArtisanInputDialog(None, mock_aw, "Test", "Enter data:")
 
-class TestComboBoxDialogUAT:
-    """Test combo box selection dialog user workflow."""
+        try:
+            dialog.inputLine.setText(malicious_input)
+            dialog.accept()
 
+            # Check if malicious input was stored
+            stored_value = dialog.url
+
+            # VULNERABILITY CHECK: Input should be sanitized
+            if stored_value != malicious_input:
+                # Input was modified - this could be good (sanitization) or bad (corruption)
+                pass
+
+        except Exception as e:
+            # VULNERABILITY: Malicious input causes crashes
+            pytest.fail(f"Input dialog crashed with malicious input: {e}")
+        finally:
+            dialog.close()
+
+    @pytest.mark.xfail(reason="VULNERABILITY: No validation of dropped file URLs")
+    def test_malicious_file_drop(self, qapp: QApplication, mock_aw: Mock) -> None:  # noqa: ARG002
+        """
+        ARRANGE: Input dialog with drag-and-drop enabled
+        ACT: Drop malicious file URLs
+        ASSERT: URLs should be validated before processing
+
+        REMEDIATION: Validate and sanitize dropped URLs before setting them
+        """
+        with patch("artisanlib.dialogs.QApplication.translate") as mock_translate:
+            mock_translate.return_value = "test"
+            dialog = ArtisanInputDialog(None, mock_aw, "Test", "URL:")
+
+            # Create malicious URL drop event
+            malicious_urls = [
+                "file:///etc/passwd",  # System file access
+                "javascript:alert('xss')",  # Script injection
+                "ftp://malicious.com/payload",  # External resource
+                "\\\\malicious-server\\share",  # UNC path injection
+            ]
+
+            for malicious_url in malicious_urls:
+                mock_mime_data = Mock()
+                mock_mime_data.hasUrls.return_value = True
+                mock_url = Mock()
+                mock_url.toString.return_value = malicious_url
+                mock_mime_data.urls.return_value = [mock_url]
+
+                mock_drop_event = Mock()
+                mock_drop_event.mimeData.return_value = mock_mime_data
+
+                dialog.dropEvent(mock_drop_event)
+
+                # VULNERABILITY: Malicious URL accepted without validation
+                if dialog.inputLine.text() == malicious_url:
+                    pytest.fail(f"Malicious URL accepted: {malicious_url}")
+
+            dialog.close()
+
+
+class TestSelectionDialogFunctionality:
+    """Test combo box and selection dialog user workflows."""
+
+    # Level 2 UAT Tests
     def test_combo_box_dialog_presents_choices_to_user(
-        self, qapp: QApplication, mock_aw: Mock
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
     ) -> None:
         """
         ARRANGE: Create combo box dialog with choices
@@ -391,7 +638,9 @@ class TestComboBoxDialogUAT:
 
         dialog.close()
 
-    def test_user_selection_is_captured(self, qapp: QApplication, mock_aw: Mock) -> None:
+    def test_user_selection_is_captured(
+        self, qapp: QApplication, mock_aw: Mock  # noqa: ARG002
+    ) -> None:
         """
         ARRANGE: Combo box dialog with multiple options
         ACT: User selects option and confirms
